fix for failing tests

+++ xugrid/regrid/structured.py
@@ -185,7 +185,7 @@ class StructuredGrid1d:
         """
         # in cases where midpoint(target) < midpoint(source), within bounding box of source,
         # neighbor will be -1. All other cases +1
-        neighbour = np.ones(source_index.size)
+        neighbour = np.ones(source_index.size,dtype=int)
         neighbour[other.midpoints[target_index] < self.midpoints[source_index]] = -1
         source_index = np.column_stack((source_index, source_index + neighbour)).ravel()
         target_index = np.repeat(target_index, 2)
@@ -227,7 +227,7 @@ class StructuredGrid1d:
         # weights = 1 - weights (since we will stack using -1 in centroids_to_linear_indexes)
         mask = other.midpoints[target_index] < self.midpoints[source_index]
         weights[mask] = 1 + weights[mask]
-        
+
         weights[weights < 0.0] = 0.0
         weights[weights > 1.0] = 1.0
         return weights
@@ -311,8 +311,8 @@ class StructuredGrid1d:
 
         source_index, target_index = self.valid_lineair_nodes_index(other)
         weights = self.compute_weights(other, source_index, target_index)
-        source_index, target_index, weights = self.centroids_to_linear_indexes(other,
-            source_index, target_index, weights
+        source_index, target_index, weights = self.centroids_to_linear_indexes(
+            other, source_index, target_index, weights
         )
         return self.sorted_output(source_index, target_index, weights)
